fix: page.tsx を Server Component ラッパーに分離してページ検出を修正
"use client" を page.tsx に直接書くと Vercel/Turbopack でルートが検出されない問題を修正。
page.tsx を Server Component ラッパー、実装を *-client.tsx に分離。
